+++ example_project/settings/base.py
@@ -35,7 +35,8 @@ USE_I18N = True
 USE_L10N = True
 
 # If you set this to False, Django will not use timezone-aware datetimes.
-USE_TZ = True
+# django-timepiece does not currently support timezones.
+USE_TZ = False
 
 # Absolute filesystem path to the directory that will hold user-uploaded files.
 # Example: '/home/media/media.lawrence.com/media/'
@@ -93,8 +94,8 @@ TEMPLATE_CONTEXT_PROCESSORS = (
     'django.contrib.messages.context_processors.messages',
     'django.core.context_processors.request',
     'timepiece.context_processors.quick_search',
-    'timepiece.context_processors.active_entries',
     'timepiece.context_processors.extra_nav',
+    'timepiece.context_processors.quick_clock_in',
 )
 
 TEMPLATE_DIRS = (
@@ -162,3 +163,8 @@ COMPRESS_PRECOMPILERS = (
     ('text/less', 'lessc {infile} {outfile}'),
 )
 COMPRESS_ROOT = '%s/static/' % PROJECT_PATH
+
+# django-timepiece settings
+TIMEPIECE_DEFAULT_LOCATION_SLUG = None
+TIMEPIECE_PROJECTS = {}
+EXTRA_NAV = {}